Link with expiry date not working (AND-13475)

Closes AND-13475

See merge request mobile/android/android!2291

+++ app/src/main/java/mega/privacy/android/app/getLink/GetLinkFragment.kt
@@ -557,7 +557,7 @@ class GetLinkFragment : BaseFragment(), DatePickerDialog.OnDateSetListener, Scro
         val date = cal.time
         val dfTimestamp = SimpleDateFormat("yyyyMMdd", Locale.getDefault())
         val dateString = dfTimestamp.format(date) + LAST_MINUTE
-        val timestamp = calculateTimestamp(dateString).toInt()
+        val timestamp = calculateTimestamp(dateString)
 
         viewModel.exportWithTimestamp(timestamp)
     }
